pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-factory"
version = "0.0.0"
description = "Personal AI factory that turns Linear tickets into GitHub PRs."
requires-python = ">=3.12"
dependencies = [
    "typer>=0.12",
    "pyyaml>=6.0",
    "requests>=2.32",
    "python-dotenv>=1.0",
    "pathspec>=1.1.1",
]

[project.scripts]
factory = "factory.cli:app"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.hatch.build.targets.wheel]
packages = ["src/factory"]

[tool.pytest.ini_options]
testpaths = ["tests"]
